pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ps-docs-rag"
version = "0.1.0"
description = "Payment Solutions documentation crawler, summarizer, and RAG Q&A system"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.85.0",
    "chromadb>=1.5.0",
    "langchain>=1.2.0",
    "langchain-community>=0.3.0",
    "langchain-chroma>=0.2.0",
    "beautifulsoup4>=4.12.0",
    "requests>=2.32.0",
    "requests-ntlm>=1.3.0",
    "python-dotenv>=1.0.0",
    "rich>=14.0.0",
    "typer>=0.12.0",
    "mcp>=1.0.0",
    "lxml>=5.0.0",
]

[project.optional-dependencies]
playwright = ["playwright>=1.44.0"]

[tool.setuptools.packages.find]
include = ["crawler*", "summarizer*", "rag*", "mcp_server*"]

[project.scripts]
ps-crawl = "crawler.crawl:app"
ps-summarize = "summarizer.summarize:app"
ps-ingest = "rag.ingest:app"
ps-query = "rag.query:app"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
